+++ licenserc.toml
@@ -44,8 +44,8 @@ excludes = [
   "cpp/apidoc",
   "cpp/thirdparty",
   "cpp/misc/cpplint.py",
-  "spark/datasources-32/src/main/scala/org/apache/graphar/datasources",
-  "spark/datasources-33/src/main/scala/org/apache/graphar/datasources",
+  "spark/datasources-32/src/main/scala/org/apache/spark/sql/graphar",
+  "spark/datasources-33/src/main/scala/org/apache/spark/sql/graphar",
   "java/src/main/java/org/apache/graphar/stdcxx/StdString.java",
   "java/src/main/java/org/apache/graphar/stdcxx/StdVector.java",
   "java/src/main/java/org/apache/graphar/stdcxx/StdSharedPtr.java",
