+++ maven-projects/spark/datasources-32/src/main/scala/org/apache/graphar/datasources/GarDataSource.scala
@@ -16,24 +16,24 @@
 
 package org.apache.graphar.datasources
 
-import scala.collection.JavaConverters._
-import scala.util.matching.Regex
-import java.util
-
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
-
-import org.apache.spark.sql.connector.catalog.{Table, TableProvider}
-import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.connector.catalog.{Table, TableProvider}
+import org.apache.spark.sql.connector.expressions.Transform
+import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.execution.datasources.csv.CSVFileFormat
 import org.apache.spark.sql.execution.datasources.orc.OrcFileFormat
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
+import org.apache.spark.sql.graphar.GarTable
+import org.apache.spark.sql.sources.DataSourceRegister
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
-import org.apache.spark.sql.sources.DataSourceRegister
-import org.apache.spark.sql.connector.expressions.Transform
+
+import java.util
+import scala.collection.JavaConverters._
+import scala.util.matching.Regex
 
 // Derived from Apache Spark 3.1.1
 // https://github.com/apache/spark/blob/1d550c4/sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/FileDataSourceV2.scala
@@ -79,8 +79,8 @@ class GarDataSource extends TableProvider with DataSourceRegister {
   }
 
   protected def getOptionsWithoutPaths(
-      map: CaseInsensitiveStringMap
-  ): CaseInsensitiveStringMap = {
+                                        map: CaseInsensitiveStringMap
+                                      ): CaseInsensitiveStringMap = {
     val withoutPath = map.asCaseSensitiveMap().asScala.filterKeys { k =>
       !k.equalsIgnoreCase("path") && !k.equalsIgnoreCase("paths")
     }
@@ -88,9 +88,9 @@ class GarDataSource extends TableProvider with DataSourceRegister {
   }
 
   protected def getTableName(
-      map: CaseInsensitiveStringMap,
-      paths: Seq[String]
-  ): String = {
+                              map: CaseInsensitiveStringMap,
+                              paths: Seq[String]
+                            ): String = {
     val hadoopConf = sparkSession.sessionState.newHadoopConfWithOptions(
       map.asCaseSensitiveMap().asScala.toMap
     )
@@ -101,9 +101,9 @@ class GarDataSource extends TableProvider with DataSourceRegister {
   }
 
   private def qualifiedPathName(
-      path: String,
-      hadoopConf: Configuration
-  ): String = {
+                                 path: String,
+                                 hadoopConf: Configuration
+                               ): String = {
     val hdfsPath = new Path(path)
     val fs = hdfsPath.getFileSystem(hadoopConf)
     hdfsPath.makeQualified(fs.getUri, fs.getWorkingDirectory).toString
@@ -149,16 +149,16 @@ class GarDataSource extends TableProvider with DataSourceRegister {
   }
 
   override def inferPartitioning(
-      options: CaseInsensitiveStringMap
-  ): Array[Transform] = {
+                                  options: CaseInsensitiveStringMap
+                                ): Array[Transform] = {
     Array.empty
   }
 
   override def getTable(
-      schema: StructType,
-      partitioning: Array[Transform],
-      properties: util.Map[String, String]
-  ): Table = {
+                         schema: StructType,
+                         partitioning: Array[Transform],
+                         properties: util.Map[String, String]
+                       ): Table = {
     // If the table is already loaded during schema inference, return it directly.
     if (t != null) {
       t
@@ -169,11 +169,11 @@ class GarDataSource extends TableProvider with DataSourceRegister {
 
   // Get the actual fall back file format.
   private def getFallbackFileFormat(
-      options: CaseInsensitiveStringMap
-  ): Class[_ <: FileFormat] = options.get("fileFormat") match {
-    case "csv"     => classOf[CSVFileFormat]
-    case "orc"     => classOf[OrcFileFormat]
+                                     options: CaseInsensitiveStringMap
+                                   ): Class[_ <: FileFormat] = options.get("fileFormat") match {
+    case "csv" => classOf[CSVFileFormat]
+    case "orc" => classOf[OrcFileFormat]
     case "parquet" => classOf[ParquetFileFormat]
-    case _         => throw new IllegalArgumentException
+    case _ => throw new IllegalArgumentException
   }
 }
