+++ maven-projects/spark/datasources-33/src/main/scala/org/apache/spark/sql/graphar/GarScan.scala
@@ -44,18 +44,18 @@ import scala.jdk.CollectionConverters._
 
 /** GarScan is a class to implement the file scan for GarDataSource. */
 case class GarScan(
-    sparkSession: SparkSession,
-    hadoopConf: Configuration,
-    fileIndex: PartitioningAwareFileIndex,
-    dataSchema: StructType,
-    readDataSchema: StructType,
-    readPartitionSchema: StructType,
-    pushedFilters: Array[Filter],
-    options: CaseInsensitiveStringMap,
-    formatName: String,
-    partitionFilters: Seq[Expression] = Seq.empty,
-    dataFilters: Seq[Expression] = Seq.empty
-) extends FileScan {
+                    sparkSession: SparkSession,
+                    hadoopConf: Configuration,
+                    fileIndex: PartitioningAwareFileIndex,
+                    dataSchema: StructType,
+                    readDataSchema: StructType,
+                    readPartitionSchema: StructType,
+                    pushedFilters: Array[Filter],
+                    options: CaseInsensitiveStringMap,
+                    formatName: String,
+                    partitionFilters: Seq[Expression] = Seq.empty,
+                    dataFilters: Seq[Expression] = Seq.empty
+                  ) extends FileScan {
 
   /** The gar format is not splitable. */
   override def isSplitable(path: Path): Boolean = false
@@ -63,8 +63,8 @@ case class GarScan(
   /** Create the reader factory according to the actual file format. */
   override def createReaderFactory(): PartitionReaderFactory =
     formatName match {
-      case "csv"     => createCSVReaderFactory()
-      case "orc"     => createOrcReaderFactory()
+      case "csv" => createCSVReaderFactory()
+      case "orc" => createOrcReaderFactory()
       case "parquet" => createParquetReaderFactory()
       case _ =>
         throw new IllegalArgumentException("Invalid format name: " + formatName)
@@ -233,9 +233,9 @@ case class GarScan(
    * chunk file in GraphAr to a single partition.
    */
   private def getFilePartitions(
-      sparkSession: SparkSession,
-      partitionedFiles: Seq[PartitionedFile]
-  ): Seq[FilePartition] = {
+                                 sparkSession: SparkSession,
+                                 partitionedFiles: Seq[PartitionedFile]
+                               ): Seq[FilePartition] = {
     val partitions = new ArrayBuffer[FilePartition]
     val currentFiles = new ArrayBuffer[PartitionedFile]
 
@@ -271,8 +271,8 @@ case class GarScan(
 
   /** Get the hash code of the object. */
   override def hashCode(): Int = formatName match {
-    case "csv"     => super.hashCode()
-    case "orc"     => getClass.hashCode()
+    case "csv" => super.hashCode()
+    case "orc" => getClass.hashCode()
     case "parquet" => getClass.hashCode()
     case _ =>
       throw new IllegalArgumentException("Invalid format name: " + formatName)
@@ -290,8 +290,8 @@ case class GarScan(
 
   /** Construct the file scan with filters. */
   def withFilters(
-      partitionFilters: Seq[Expression],
-      dataFilters: Seq[Expression]
-  ): FileScan =
+                   partitionFilters: Seq[Expression],
+                   dataFilters: Seq[Expression]
+                 ): FileScan =
     this.copy(partitionFilters = partitionFilters, dataFilters = dataFilters)
 }
