+++ maven-projects/spark/datasources-33/src/main/scala/org/apache/spark/sql/graphar/GarWriteBuilder.scala
@@ -27,7 +27,6 @@ import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
 import org.apache.hadoop.mapreduce.Job
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat
-import org.apache.hadoop.mapreduce.Job
 
 import org.apache.spark.sql.execution.datasources.OutputWriterFactory
 import org.apache.spark.sql.SparkSession
@@ -41,7 +40,6 @@ import org.apache.spark.sql.connector.write.{
 import org.apache.spark.sql.execution.datasources.{
   BasicWriteJobStatsTracker,
   DataSource,
-  OutputWriterFactory,
   WriteJobDescription
 }
 import org.apache.spark.sql.execution.metric.SQLMetric
@@ -52,11 +50,11 @@ import org.apache.spark.sql.execution.datasources.v2.FileBatchWrite
 import org.apache.spark.sql.catalyst.expressions.AttributeReference
 
 abstract class GarWriteBuilder(
-    paths: Seq[String],
-    formatName: String,
-    supportsDataType: DataType => Boolean,
-    info: LogicalWriteInfo
-) extends WriteBuilder {
+                                paths: Seq[String],
+                                formatName: String,
+                                supportsDataType: DataType => Boolean,
+                                info: LogicalWriteInfo
+                              ) extends WriteBuilder {
   private val schema = info.schema()
   private val queryId = info.queryId()
   private val options = info.options()
@@ -90,11 +88,11 @@ abstract class GarWriteBuilder(
   }
 
   def prepareWrite(
-      sqlConf: SQLConf,
-      job: Job,
-      options: Map[String, String],
-      dataSchema: StructType
-  ): OutputWriterFactory
+                    sqlConf: SQLConf,
+                    job: Job,
+                    options: Map[String, String],
+                    dataSchema: StructType
+                  ): OutputWriterFactory
 
   private def validateInputs(caseSensitiveAnalysis: Boolean): Unit = {
     assert(schema != null, "Missing input data schema")
@@ -127,12 +125,12 @@ abstract class GarWriteBuilder(
   }
 
   private def createWriteJobDescription(
-      sparkSession: SparkSession,
-      hadoopConf: Configuration,
-      job: Job,
-      pathName: String,
-      options: Map[String, String]
-  ): WriteJobDescription = {
+                                         sparkSession: SparkSession,
+                                         hadoopConf: Configuration,
+                                         job: Job,
+                                         pathName: String,
+                                         options: Map[String, String]
+                                       ): WriteJobDescription = {
     val caseInsensitiveOptions = CaseInsensitiveMap(options)
     // Note: prepareWrite has side effect. It sets "job".
     val outputWriterFactory =
